backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pydantic
python-dotenv
gunicorn
